chore: drop u-boot patches
Drop the u-boot patches which allowed building u-boot flashable to SPI.
The board wiki already has links to flashing their own u-boot to SPI.
It's a pain to update u-boot for us due to patches. The installation
docs for the board will be updated to point to the board wiki for
flashing u-boot for SPI. This also reduces our maintenance work and
makes it easier to update u-boot.

Signed-off-by: Noel Georgi <git@frezbo.dev>