fix: skip loading /dtb when booting from eMMC
This commit updates U-Boot to 2024.10 in order to fix booting from eMMC
without the kernel DTB. U-Boot 2024.04 and 2404.07  ignore the DTB on
the EFI partition when booting from eMMC.

This is not an issue when booting from NVMe (and having a U-Boot spi
image on eMMC). Since I boot from NVMe I failed to notice getting the
Turing RK1 upstream in a very narrow timeframe. Thanks to @rducom for
bringing this problem to my attention in very short notice.

As an upgrade bonus, KASLR is enabled out of the box, meaning I do not
need my U-Boot patch anymore to forward entropy to the kernel.

Signed-off-by: Nico Berlee <nico.berlee@on2it.net>
Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>